fix: create logger
