+++ packages/features/bookings/lib/handleNewBooking/logger.ts
@@ -1,0 +1,11 @@
+import logger from "@calcom/lib/logger";
+
+export const createLoggerWithEventDetails = (
+  eventTypeId: number,
+  reqBodyUser: string | string[] | undefined,
+  eventTypeSlug: string | undefined
+) => {
+  return logger.getSubLogger({
+    prefix: ["book:user", `${eventTypeId}:${reqBodyUser}/${eventTypeSlug}`],
+  });
+};
