+++ packages/features/bookings/lib/handleSeats/reschedule/owner/ownerRescheduleSeatedBooking.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-restricted-imports
 import type EventManager from "@calcom/lib/EventManager";
 
-import type { createLoggerWithEventDetails } from "../../../handleNewBooking";
+import type { createLoggerWithEventDetails } from "../../../handleNewBooking/logger";
 import type {
   NewTimeSlotBooking,
   SeatedBooking,
